pyproject.toml
[project]
name = "windy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.3",
    "django-ninja>=1.4.3",
    "django-rainbowtests>=0.6.0",
    "django-taggit>=6.1.0",
    "geojson-pydantic>=2.0.0",
    "psycopg2-binary>=2.9.10",
]
